Rename server getSamData.js utility file to bap.js
